Refactor OTel TLS syscall tester helpers
